prepare BottomTabBarFloatingActionButton
