+++ src/pages/home/sidebar/BottomTabBarFloatingActionButton/index.native.tsx
@@ -1,0 +1,3 @@
+import FloatingActionButton from '@components/FloatingActionButton';
+
+export default FloatingActionButton;
