+++ src/pages/home/sidebar/BottomTabBarFloatingActionButton/types.ts
@@ -1,0 +1,5 @@
+type FloatingActionButtonPopoverMenuRef = {
+    hideCreateMenu: () => void;
+};
+
+export default FloatingActionButtonPopoverMenuRef;
