+++ src/pages/home/sidebar/SidebarScreen/index.js
@@ -1,50 +1,18 @@
-import React, {useCallback, useRef} from 'react';
+import React from 'react';
 import useWindowDimensions from '@hooks/useWindowDimensions';
 import FreezeWrapper from '@libs/Navigation/FreezeWrapper';
 import BaseSidebarScreen from './BaseSidebarScreen';
-import FloatingActionButtonAndPopover from './FloatingActionButtonAndPopover';
 import sidebarPropTypes from './sidebarPropTypes';
 
 function SidebarScreen(props) {
-    const popoverModal = useRef(null);
     const {isSmallScreenWidth} = useWindowDimensions();
-
-    /**
-     * Method to hide popover when dragover.
-     */
-    const hidePopoverOnDragOver = useCallback(() => {
-        if (!popoverModal.current) {
-            return;
-        }
-        popoverModal.current.hideCreateMenu();
-    }, []);
-
-    /**
-     * Method create event listener
-     */
-    const createDragoverListener = () => {
-        document.addEventListener('dragover', hidePopoverOnDragOver);
-    };
-
-    /**
-     * Method remove event listener.
-     */
-    const removeDragoverListener = () => {
-        document.removeEventListener('dragover', hidePopoverOnDragOver);
-    };
 
     return (
         <FreezeWrapper keepVisible={!isSmallScreenWidth}>
             <BaseSidebarScreen
                 // eslint-disable-next-line react/jsx-props-no-spreading
                 {...props}
-            >
-                <FloatingActionButtonAndPopover
-                    ref={popoverModal}
-                    onShowCreateMenu={createDragoverListener}
-                    onHideCreateMenu={removeDragoverListener}
-                />
-            </BaseSidebarScreen>
+            />
         </FreezeWrapper>
     );
 }
